+++ CHANGELOG.md
@@ -9,3 +9,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * basic testsuite
 * simple html-report from cli-run (xml-based)
 * testsuite is running at ci.ilias.de
+
+## develop-branch
+### Added
+* bin/script/ilias/reset_client.sh (reset one client with another one)
+
+### Removed
+* bin/script/update_ilias.sh (use this instead: https://svnext.databay.de/scm/git/PerfTestsuiteHelper)